Share cloud sync test helpers
